+++ gazebo/gui/building/BuildingMaker.cc
@@ -33,6 +33,7 @@
 #include "gazebo/gui/GuiIface.hh"
 #include "gazebo/gui/EntityMaker.hh"
 
+#include "gazebo/common/SystemPaths.hh"
 #ifdef HAVE_GTS
   #include "gazebo/common/Mesh.hh"
   #include "gazebo/common/MeshManager.hh"
@@ -52,9 +53,78 @@ using namespace gui;
 double BuildingMaker::conversionScale;
 
 /////////////////////////////////////////////////
+// Helper function to generate a valid folder name from a human-readable model
+// name.
+std::string GetFolderNameFromModelName(const std::string &_modelName)
+{
+  // Auto-generate folder name based on model name
+  std::string foldername = _modelName;
+
+  std::vector<std::pair<std::string, std::string> > replacePairs;
+  replacePairs.push_back(std::pair<std::string, std::string>(" ", "_"));
+
+  for (unsigned int i = 0; i < replacePairs.size(); ++i)
+  {
+    std::string forbiddenChar = replacePairs[i].first;
+    std::string replaceChar = replacePairs[i].second;
+    size_t index = foldername.find(forbiddenChar);
+    while (index != std::string::npos)
+    {
+      foldername.replace(index, forbiddenChar.size(), replaceChar);
+      index = foldername.find(forbiddenChar);
+    }
+  }
+
+  return foldername;
+}
+
+/////////////////////////////////////////////////
+// Add the parent folder of _path to the model path represented by SystemPaths,
+// notify InsertModelWidget to display the model name in the "Insert Models"
+// tab, and write the parent folder filename to gui.ini
+void AddDirToModelPaths(const std::string& _path)
+{
+  std::string parentDirectory = boost::filesystem::path(_path)
+                                  .parent_path().string();
+
+  std::list<std::string> modelPaths =
+              gazebo::common::SystemPaths::Instance()->GetModelPaths();
+  std::list<std::string>::iterator iter;
+  for (iter = modelPaths.begin();
+       iter != modelPaths.end(); ++iter)
+  {
+    if (iter->compare(parentDirectory) == 0)
+    {
+      break;
+    }
+  }
+
+  gazebo::common::SystemPaths::Instance()->
+    AddModelPathsUpdate(parentDirectory);
+
+  std::string additionalProperties =
+    gui::getINIProperty<std::string>("model_paths.filenames", "");
+  if (additionalProperties.find(parentDirectory) == std::string::npos)
+  {
+    // Add it to gui.ini
+    gui::setINIProperty("model_paths.filenames", parentDirectory);
+
+    // Save any changes that were made to the property tree
+    // TODO: check gui.ini env variable
+    char *home = getenv("HOME");
+    if (home)
+    {
+      boost::filesystem::path guiINIPath = home;
+      guiINIPath  = guiINIPath / ".gazebo" / "gui.ini";
+      saveINI(guiINIPath);
+    }
+  }
+}
+
+/////////////////////////////////////////////////
 BuildingMaker::BuildingMaker() : EntityMaker()
 {
-  this->buildingDefaultName = "BuildingDefaultName";
+  this->buildingDefaultName = "Untitled";
   this->modelName = this->buildingDefaultName;
 
   this->conversionScale = 0.01;
@@ -73,25 +143,26 @@ BuildingMaker::BuildingMaker() : EntityMaker()
   this->modelTemplateSDF->SetFromString(this->GetTemplateSDFString());
 
   this->connections.push_back(
-  gui::editor::Events::ConnectSaveBuildingEditor(
-    boost::bind(&BuildingMaker::OnSave, this, _1)));
+    gui::editor::Events::ConnectSaveBuildingEditor(
+      boost::bind(&BuildingMaker::OnSave, this, _1)));
   this->connections.push_back(
-  gui::editor::Events::ConnectDiscardBuildingEditor(
-    boost::bind(&BuildingMaker::OnDiscard, this)));
+    gui::editor::Events::ConnectSaveAsBuildingEditor(
+      boost::bind(&BuildingMaker::OnSaveAs, this, _1)));
   this->connections.push_back(
-  gui::editor::Events::ConnectDoneBuildingEditor(
-    boost::bind(&BuildingMaker::OnDone, this, _1)));
+    gui::editor::Events::ConnectNewBuildingEditor(
+      boost::bind(&BuildingMaker::OnNew, this)));
   this->connections.push_back(
-  gui::editor::Events::ConnectExitBuildingEditor(
-    boost::bind(&BuildingMaker::OnExit, this)));
+    gui::editor::Events::ConnectExitBuildingEditor(
+      boost::bind(&BuildingMaker::OnExit, this)));
+  this->connections.push_back(
+    gui::editor::Events::ConnectBuildingNameChanged(
+      boost::bind(&BuildingMaker::OnNameChanged, this, _1)));
   this->connections.push_back(
   gui::editor::Events::ConnectChangeBuildingLevel(
     boost::bind(&BuildingMaker::OnChangeLevel, this, _1)));
 
   this->saveDialog =
       new FinishBuildingDialog(FinishBuildingDialog::MODEL_SAVE, 0);
-  this->finishDialog =
-      new FinishBuildingDialog(FinishBuildingDialog::MODEL_FINISH, 0);
 }
 
 /////////////////////////////////////////////////
@@ -100,8 +171,6 @@ BuildingMaker::~BuildingMaker()
 //  this->camera.reset();
   if (this->saveDialog)
     delete this->saveDialog;
-  if (this->finishDialog)
-    delete this->finishDialog;
 }
 
 /////////////////////////////////////////////////
@@ -276,6 +345,7 @@ std::string BuildingMaker::AddWall(const QVector3D &_size,
   this->allItems[linkName] = wallManip;
 
   linkVisual->SetVisibilityFlags(GZ_VISIBILITY_GUI);
+  this->BuildingChanged();
   return linkName;
 }
 
@@ -303,9 +373,8 @@ std::string BuildingMaker::AddWindow(const QVector3D &_size,
 
   sdf::ElementPtr visualElem =  this->modelTemplateSDF->root
       ->GetElement("model")->GetElement("link")->GetElement("visual");
-  visualElem->GetElement("material")->ClearElements();
-  visualElem->GetElement("material")->AddElement("ambient")
-      ->Set(gazebo::common::Color(0, 0, 1));
+  visualElem->GetElement("material")->GetElement("script")->GetElement("name")
+      ->Set("Gazebo/BuildingFrame");
   visualElem->AddElement("cast_shadows")->Set(false);
   visVisual->Load(visualElem);
 
@@ -321,6 +390,7 @@ std::string BuildingMaker::AddWindow(const QVector3D &_size,
   this->allItems[linkName] = windowManip;
 
   linkVisual->SetVisibilityFlags(GZ_VISIBILITY_GUI);
+  this->BuildingChanged();
   return linkName;
 }
 
@@ -349,9 +419,8 @@ std::string BuildingMaker::AddDoor(const QVector3D &_size,
 
   sdf::ElementPtr visualElem =  this->modelTemplateSDF->root
       ->GetElement("model")->GetElement("link")->GetElement("visual");
-  visualElem->GetElement("material")->ClearElements();
-  visualElem->GetElement("material")->AddElement("ambient")
-      ->Set(gazebo::common::Color(1, 1, 0));
+  visualElem->GetElement("material")->GetElement("script")->GetElement("name")
+      ->Set("Gazebo/BuildingFrame");
   visualElem->AddElement("cast_shadows")->Set(false);
   visVisual->Load(visualElem);
 
@@ -367,6 +436,7 @@ std::string BuildingMaker::AddDoor(const QVector3D &_size,
   this->allItems[linkName] = doorManip;
 
   linkVisual->SetVisibilityFlags(GZ_VISIBILITY_GUI);
+  this->BuildingChanged();
   return linkName;
 }
 
@@ -439,6 +509,8 @@ std::string BuildingMaker::AddStairs(const QVector3D &_size,
   }
 
   linkVisual->SetVisibilityFlags(GZ_VISIBILITY_GUI);
+  this->BuildingChanged();
+
   return linkName;
 }
 
@@ -485,6 +557,7 @@ std::string BuildingMaker::AddFloor(const QVector3D &_size,
   this->allItems[linkName] = floorManip;
 
   linkVisual->SetVisibilityFlags(GZ_VISIBILITY_GUI);
+  this->BuildingChanged();
   return linkName;
 }
 
@@ -505,6 +578,7 @@ void BuildingMaker::RemovePart(const std::string &_partName)
     scene->RemoveVisual(visParent);
   this->allItems.erase(_partName);
   delete manip;
+  this->BuildingChanged();
 }
 
 /////////////////////////////////////////////////
@@ -540,12 +614,21 @@ void BuildingMaker::Reset()
 
   rendering::ScenePtr scene = gui::get_active_camera()->GetScene();
 
+  if (!scene)
+  {
+    gzerr << "Couldn't get scene node from BuildingMaker" << std::endl;
+    return;
+  }
+
   if (this->modelVisual)
     scene->RemoveVisual(this->modelVisual);
 
-  this->saved = false;
-  this->saveLocation = QDir::homePath().toStdString();
-  this->modelName = this->buildingDefaultName;
+  this->currentSaveState = NEVER_SAVED;
+  this->SetModelName(this->buildingDefaultName);
+  this->defaultPath = (QDir::homePath() + "/building_editor_models")
+                        .toStdString();
+  this->saveLocation = defaultPath + "/" +
+                        GetFolderNameFromModelName(this->modelName);
 
   this->modelVisual.reset(new rendering::Visual(this->modelName,
       scene->GetWorldVisual()));
@@ -572,19 +655,28 @@ bool BuildingMaker::IsActive() const
 void BuildingMaker::SetModelName(const std::string &_modelName)
 {
   this->modelName = _modelName;
+  this->saveDialog->SetModelName(_modelName);
+  this->BuildingChanged();
+  // send to palette?
 }
 
 /////////////////////////////////////////////////
 void BuildingMaker::SaveToSDF(const std::string &_savePath)
 {
-  this->saveLocation = _savePath;
   std::ofstream savefile;
-  boost::filesystem::path path;
-  path = boost::filesystem::operator/(this->saveLocation,
-      this->modelName + ".sdf");
+  boost::filesystem::path path(_savePath);
+  path = path / "model.sdf";
+
+  // FIXME
   savefile.open(path.string().c_str());
+  if (!savefile.is_open())
+  {
+    gzerr << "Couldn't open file for writing: " << path.string() << std::endl;
+    return;
+  }
   savefile << this->modelSDF->ToString();
   savefile.close();
+  gzdbg << "Saved file to " << path.string() << std::endl;
 }
 
 /////////////////////////////////////////////////
@@ -626,7 +718,8 @@ void BuildingMaker::GenerateSDF()
   std::stringstream visualNameStream;
   std::stringstream collisionNameStream;
 
-  modelElem->GetAttribute("name")->Set(this->modelName);
+  modelElem->GetAttribute("name")->Set(
+    GetFolderNameFromModelName(this->modelName));
 
   std::map<std::string, BuildingModelManip *>::iterator itemsIt;
 
@@ -864,6 +957,8 @@ void BuildingMaker::GenerateSDF()
               GetElement("size")->Set(visual->GetScale());
           visualElem->GetElement("material")->GetElement("ambient")->
               Set(buildingModelManip->GetColor());
+          visualElem->GetElement("material")->GetElement("script")
+              ->GetElement("name")->Set(buildingModelManip->GetTexture());
         }
       }
     }
@@ -1096,8 +1191,27 @@ void BuildingMaker::GenerateSDFWithCSG()
 /////////////////////////////////////////////////
 void BuildingMaker::CreateTheEntity()
 {
-  this->GenerateSDF();
   msgs::Factory msg;
+  // Create a new name if the model exists
+  if (!this->modelSDF->root->HasElement("model"))
+  {
+    gzerr << "Generated invalid SDF! Cannot create entity." << std::endl;
+    return;
+  }
+
+  sdf::ElementPtr modelElem = this->modelSDF->root->GetElement("model");
+  std::string modelElemName = modelElem->Get<std::string>("name");
+  if (has_entity_name(modelElemName))
+  {
+    int i = 0;
+    while (has_entity_name(modelElemName))
+    {
+      modelElemName = modelElem->Get<std::string>("name") + "_" +
+        boost::lexical_cast<std::string>(i++);
+    }
+    modelElem->GetAttribute("name")->Set(modelElemName);
+  }
+
   msg.set_sdf(this->modelSDF->ToString());
   this->makerPub->Publish(msg);
 }
@@ -1144,6 +1258,23 @@ double BuildingMaker::Convert(double _value)
 double BuildingMaker::ConvertAngle(double _angle)
 {
   return GZ_DTOR(_angle);
+}
+
+std::string BuildingMaker::GetTemplateConfigString()
+{
+  std::ostringstream newModelStr;
+  newModelStr << "<?xml version=\"1.0\"?>"
+  << "<model>"
+  <<   "<name>building_template_model</name>"
+  <<   "<version>1.0</version>"
+  <<   "<sdf version=\"1.5\">model.sdf</sdf>"
+  <<   "<author>"
+  <<     "<name>author_name</name>"
+  <<     "<email>author_email</email>"
+  <<   "</author>"
+  <<   "<description>Made with the Gazebo Building Editor</description>"
+  << "</model>";
+  return newModelStr.str();
 }
 
 /////////////////////////////////////////////////
@@ -1393,76 +1524,287 @@ void BuildingMaker::SubdivideRectSurface(const QRectF &_surface,
 }
 
 /////////////////////////////////////////////////
-void BuildingMaker::OnDiscard()
+void BuildingMaker::OnNew()
 {
-  int ret = QMessageBox::warning(0, QString("Discard"),
-      QString("Are you sure you want to discard\n"
-      "your model? All of your work will\n"
-      "be lost."),
-      QMessageBox::Yes | QMessageBox::Cancel,
-      QMessageBox::Cancel);
-
-  switch (ret)
+  if (this->allItems.empty())
   {
-    case QMessageBox::Yes:
-      this->Reset();
-      gui::editor::Events::discardBuildingModel();
+    this->Reset();
+    gui::editor::Events::newBuildingModel();
+    return;
+  }
+  QString msg;
+  QMessageBox msgBox(QMessageBox::Warning, QString("New"), msg);
+  QPushButton *cancelButton = msgBox.addButton("Cancel", QMessageBox::YesRole);
+  QPushButton *saveButton = msgBox.addButton("Save", QMessageBox::YesRole);
+
+  switch (this->currentSaveState)
+  {
+    case ALL_SAVED:
+    {
+      msg.append("Are you sure you want to close this model and open a new "
+                 "canvas?\n\n");
+      msgBox.addButton("New Canvas", QMessageBox::ApplyRole);
+      saveButton->hide();
       break;
-    case QMessageBox::Cancel:
-    // Do nothing
-    break;
+    }
+    case UNSAVED_CHANGES:
+    case NEVER_SAVED:
+    {
+      msg.append("You have unsaved changes. Do you want to save this model "
+                 "and open a new canvas?\n\n");
+      msgBox.addButton("Don't Save", QMessageBox::ApplyRole);
+      break;
+    }
     default:
-    break;
+      return;
+  }
+
+  msg.append("Once you open a new canvas, your current model will no longer "
+             "be editable.");
+  msgBox.setText(msg);
+
+  msgBox.exec();
+
+  if (msgBox.clickedButton() != cancelButton)
+  {
+    if (msgBox.clickedButton() == saveButton)
+    {
+      if (!this->OnSave(this->modelName))
+      {
+        return;
+      }
+    }
+
+    this->Reset();
+    gui::editor::Events::newBuildingModel();
   }
 }
 
-/////////////////////////////////////////////////
-void BuildingMaker::OnSave(const std::string &_saveName)
+void BuildingMaker::SaveModelFiles()
 {
-  if (_saveName != "")
-    this->SetModelName(_saveName);
+  this->SetModelName(this->modelName);
+  this->GenerateConfig();
+  this->SaveToConfig(this->saveLocation);
+  this->GenerateSDF();
+  this->SaveToSDF(this->saveLocation);
+  this->currentSaveState = ALL_SAVED;
+}
 
-  if (this->saved)
+/////////////////////////////////////////////////
+void BuildingMaker::GenerateConfig()
+{
+  // Create an xml config file
+  this->modelConfig.Clear();
+  this->modelConfig.Parse(this->GetTemplateConfigString().c_str());
+
+  TiXmlElement *modelXML = this->modelConfig.FirstChildElement("model");
+  if (!modelXML)
   {
-    this->GenerateSDF();
-    this->SaveToSDF(this->saveLocation);
+    gzerr << "No model name in default config file\n";
+    return;
+  }
+  TiXmlElement *modelNameXML = modelXML->FirstChildElement("name");
+  modelNameXML->FirstChild()->SetValue(this->modelName);
+
+  TiXmlElement *versionXML = modelXML->FirstChildElement("version");
+  if (!versionXML)
+  {
+    gzerr << "Couldn't find model version" << std::endl;
+    versionXML->FirstChild()->SetValue("1.0");
   }
   else
   {
-    this->saveDialog->SetModelName(this->modelName);
-    this->saveDialog->SetSaveLocation(this->saveLocation);
-    if (this->saveDialog->exec() == QDialog::Accepted)
+    versionXML->FirstChild()->SetValue(this->version);
+  }
+
+  TiXmlElement *descriptionXML = modelXML->FirstChildElement("description");
+  if (!descriptionXML)
+  {
+    gzerr << "Couldn't find model description" << std::endl;
+    descriptionXML->FirstChild()->SetValue("");
+  }
+  else
+  {
+    descriptionXML->FirstChild()->SetValue(this->description);
+  }
+
+  // TODO: Multiple authors
+  TiXmlElement *authorXML = modelXML->FirstChildElement("author");
+  if (!authorXML)
+  {
+    gzerr << "Couldn't find model author" << std::endl;
+  }
+  else
+  {
+    TiXmlElement *authorChild = authorXML->FirstChildElement("name");
+    if (!authorChild)
     {
-      this->SetModelName(this->saveDialog->GetModelName());
-      this->saveLocation = this->saveDialog->GetSaveLocation();
-      this->GenerateSDF();
-      this->SaveToSDF(this->saveLocation);
-      this->saved = true;
-      // Send confirmation that model has been saved
-      gui::editor::Events::saveBuildingModel(this->modelName,
-          this->saveLocation);
+      gzerr << "Couldn't find author name" << std::endl;
+      authorChild->FirstChild()->SetValue("");
+    }
+    else
+    {
+      authorChild->FirstChild()->SetValue(this->authorName);
+    }
+    authorChild = authorXML->FirstChildElement("email");
+    if (!authorChild)
+    {
+      gzerr << "Couldn't find author email" << std::endl;
+      authorChild->FirstChild()->SetValue("");
+    }
+    else
+    {
+      authorChild->FirstChild()->SetValue(this->authorEmail);
     }
   }
 }
 
 /////////////////////////////////////////////////
-void BuildingMaker::OnDone(const std::string &_saveName)
+void BuildingMaker::SaveToConfig(const std::string &_savePath)
+{
+  boost::filesystem::path path(_savePath);
+  path = path / "model.config";
+  const char* modelConfigString = path.string().c_str();
+
+  this->modelConfig.SaveFile(modelConfigString);
+  gzdbg << "Saved file to " << modelConfigString << std::endl;
+}
+
+/////////////////////////////////////////////////
+bool BuildingMaker::OnSave(const std::string &_saveName)
 {
   if (_saveName != "")
     this->SetModelName(_saveName);
 
-  this->finishDialog->SetModelName(this->modelName);
-  this->finishDialog->SetSaveLocation(this->saveLocation);
-  if (this->finishDialog->exec() == QDialog::Accepted)
+  switch (this->currentSaveState)
   {
-    this->SetModelName(this->finishDialog->GetModelName());
-    this->saveLocation = this->finishDialog->GetSaveLocation();
-    this->GenerateSDF();
-    this->SaveToSDF(this->saveLocation);
-    this->FinishModel();
-    gui::editor::Events::discardBuildingModel();
-    gui::editor::Events::finishBuildingModel();
+    case UNSAVED_CHANGES:
+    {
+      // TODO: Subtle filesystem race condition
+      this->SaveModelFiles();
+      AddDirToModelPaths(this->saveLocation);
+      gui::editor::Events::saveBuildingModel(this->modelName,
+          this->saveLocation);
+      return true;
+    }
+    case NEVER_SAVED:
+    {
+      return this->OnSaveAs(_saveName);
+    }
+    default:
+      return false;
   }
+}
+
+/////////////////////////////////////////////////
+bool BuildingMaker::OnSaveAs(const std::string &_saveName)
+{
+  this->saveDialog->SetModelName(_saveName);
+
+  if (this->saveLocation.length() > 0)
+  {
+    this->saveDialog->SetSaveLocation(this->saveLocation);
+  }
+  if (this->saveDialog->exec() == QDialog::Accepted)
+  {
+    if (this->saveDialog->GetModelName().size() == 0)
+    {
+      QMessageBox msgBox(QMessageBox::Warning, QString("Empty Name"),
+                       QString("Please give your model a non-empty name."));
+
+      msgBox.exec();
+      return this->OnSaveAs(_saveName);
+    }
+    if (this->saveDialog->GetSaveLocation().size() == 0)
+    {
+      QMessageBox msgBox(QMessageBox::Warning, QString("Empty Location"),
+             QString("Please give a path to where your model will be saved."));
+
+      msgBox.exec();
+      return this->OnSaveAs(_saveName);
+    }
+
+    this->modelName = this->saveDialog->GetModelName();
+    this->saveLocation = this->saveDialog->GetSaveLocation();
+    this->authorName = this->saveDialog->GetAuthorName();
+    this->authorEmail = this->saveDialog->GetAuthorEmail();
+    this->description = this->saveDialog->GetDescription();
+    this->version = this->saveDialog->GetVersion();
+
+    if (this->modelName.compare(this->buildingDefaultName) == 0)
+    {
+      // Parse saveLocation and set model name
+      boost::filesystem::path saveLocPath(this->saveLocation);
+      this->SetModelName(saveLocPath.filename().string());
+    }
+
+    boost::filesystem::path path;
+    path = path / this->saveLocation;
+    if (!boost::filesystem::exists(path))
+    {
+      if (!boost::filesystem::create_directories(path))
+      {
+        gzerr << "Couldn't create folder for model files." << std::endl;
+        return false;
+      }
+      gzmsg << "Created folder " << path << " for model files." << std::endl;
+    }
+
+    boost::filesystem::path modelConfigPath = path / "model.config";
+
+    boost::filesystem::path sdfPath = path / "model.sdf";
+
+    // Before writing
+    if (boost::filesystem::exists(sdfPath) ||
+          boost::filesystem::exists(modelConfigPath))
+    {
+      std::string msg = "A model named " + this->modelName +
+                        " already exists in folder " + path.string() + ".\n\n"
+                        "Do you wish to overwrite the existing model files?\n";
+
+      QMessageBox msgBox(QMessageBox::Warning, QString("Files Exist"),
+                         QString(msg.c_str()));
+
+      QPushButton *saveButton = msgBox.addButton("Save",
+                                                 QMessageBox::ApplyRole);
+      msgBox.addButton(QMessageBox::Cancel);
+      msgBox.exec();
+      if (msgBox.clickedButton() != saveButton)
+      {
+        return this->OnSaveAs(this->modelName);
+      }
+    }
+
+    this->SaveModelFiles();
+
+    AddDirToModelPaths(this->saveLocation);
+
+    gui::editor::Events::saveBuildingModel(this->modelName, this->saveLocation);
+    return true;
+  }
+  return false;
+}
+
+/////////////////////////////////////////////////
+void BuildingMaker::OnNameChanged(const std::string &_name)
+{
+  if (_name.compare(this->modelName) == 0)
+  {
+    return;
+  }
+  this->SetModelName(_name);
+
+  if (this->currentSaveState == NEVER_SAVED)
+  {
+    // Set new saveLocation
+    boost::filesystem::path oldPath(this->saveLocation);
+
+    boost::filesystem::path newPath = oldPath.parent_path() /
+          GetFolderNameFromModelName(_name);
+    this->saveLocation = newPath.string();
+  }
+
+  this->BuildingChanged();
 }
 
 /////////////////////////////////////////////////
@@ -1470,38 +1812,78 @@ void BuildingMaker::OnExit()
 {
   if (this->allItems.empty())
   {
+    this->Reset();
     gui::editor::Events::finishBuildingModel();
     return;
   }
 
-  QMessageBox msgBox;
-  msgBox.setWindowTitle("Exit");
-  msgBox.setText("Save changes before exiting? If you do not\n"
-        "save, all of your work will be lost!\n\n"
-        "Note: Once you exit the Building Editor, your\n"
-        "building will no longer be editable.");
-  QPushButton *discardButton = msgBox.addButton("Don't Save, Exit",
-      QMessageBox::ActionRole);
-  msgBox.addButton("Cancel", QMessageBox::ActionRole);
-  QPushButton *doneButton = msgBox.addButton("Save", QMessageBox::ActionRole);
-  msgBox.setDefaultButton(doneButton);
-
-
-  msgBox.exec();
-  if (msgBox.clickedButton() == doneButton)
+  switch (this->currentSaveState)
   {
-    this->OnDone();
+    case ALL_SAVED:
+    {
+      QString msg("Once you exit the Building Editor, "
+      "your building will no longer be editable.\n\n"
+      "Are you ready to exit?\n\n");
+      QMessageBox msgBox(QMessageBox::NoIcon, QString("Exit"), msg);
+      msgBox.addButton("Exit", QMessageBox::ApplyRole);
+      QPushButton *cancelButton = msgBox.addButton(QMessageBox::Cancel);
+      msgBox.exec();
+      if (msgBox.clickedButton() == cancelButton)
+      {
+        return;
+      }
+      this->FinishModel();
+      break;
+    }
+    case UNSAVED_CHANGES:
+    case NEVER_SAVED:
+    {
+      QString msg("Save Changes before exiting?\n\n"
+          "Note: Once you exit the Building Editor, "
+          "your building will no longer be editable.\n\n");
+
+      QMessageBox msgBox(QMessageBox::NoIcon, QString("Exit"), msg);
+      QPushButton *cancelButton = msgBox.addButton("Cancel",
+          QMessageBox::ApplyRole);
+      QPushButton *saveButton = msgBox.addButton("Save and Exit",
+          QMessageBox::ApplyRole);
+      msgBox.addButton("Don't Save, Exit", QMessageBox::ApplyRole);
+      msgBox.exec();
+      if (msgBox.clickedButton() == cancelButton)
+        return;
+
+      if (msgBox.clickedButton() == saveButton)
+      {
+        if (!this->OnSave(this->modelName))
+        {
+          return;
+        }
+      }
+      break;
+    }
+    default:
+      return;
   }
-  else if (msgBox.clickedButton() == discardButton)
-  {
-    this->Reset();
-    gui::editor::Events::discardBuildingModel();
-    gui::editor::Events::finishBuildingModel();
-  }
+
+  // Create entity on main window up to the saved point
+  if (this->currentSaveState != NEVER_SAVED)
+    this->FinishModel();
+
+  this->Reset();
+
+  gui::editor::Events::newBuildingModel();
+  gui::editor::Events::finishBuildingModel();
 }
 
 /////////////////////////////////////////////////
 void BuildingMaker::OnChangeLevel(int _level)
 {
   this->currentLevel = _level;
+}
+
+/////////////////////////////////////////////////
+void BuildingMaker::BuildingChanged()
+{
+  if (this->currentSaveState != NEVER_SAVED)
+    this->currentSaveState = UNSAVED_CHANGES;
 }